moving k_fold_split_copy into whoot package
