+++ create_dataset/create_folds/strat_k_folds.py
@@ -15,7 +15,7 @@ import pandas as pd
 import numpy as np
 
 
-from k_fold_split_copy import solve
+from whoot.k_fold_split_copy import solve
 
 
 def create_strat_folds(df, site_ids):  # pylint: disable-msg=too-many-locals
