+++ whoot/__init__.py
@@ -1,3 +1,4 @@
 __version__ = "0.1.1.dev0"
 
 from .audio_utils import expand_window
+from .k_fold_split_copy import solve
